pyproject.toml
[project]
name = "artistic-intelligence-data"
version = "0.1.0"
description = "Data collector, query API and more to create digital art."
readme = "README.md"
authors = [
    {name = "Robin Klaassen", email = "info@robinklaassen.com"},
]
license = {text = "GPL-3.0-or-later"}
requires-python = "<4.0,>=3.12"
dependencies = [
#    "APScheduler",
    "connectorx",
    "fastapi[standard]",
    "geopandas",  # TODO move to scripting dependency
#    "gunicorn",
    "influxdb-client[ciso]",
    "pandas",  # TODO remove from core in favor of polars
    "polars",
    "psycopg[binary]",  # TODO remove in favor of connectorx
    "pyarrow",
    "pydantic",
    "python-dotenv",
#    "requests",
    "structlog",
    "uvicorn[standard]",
]

[project.urls]
homepage = "https://artisticintelligence.nl"
repository = "https://github.com/robinklaassen/artistic-intelligence-data"

[build-system]
requires = ["uv_build"]
build-backend = "uv_build"

[tool.poe.tasks]
lint = "ruff check"
lint-fix = "ruff check --fix"
format = "ruff format --check"
format-fix = "ruff format"
type = "ty check"
test = "pytest tests"
sec = "bandit -r src"
all = ["lint-fix", "format-fix", "type", "test", "sec"]

[tool.ruff]
extend-exclude = ["*.ipynb"]
line-length = 120

[tool.ruff.lint]
select = [
    "E",
    "F",
    "UP",
    "B",
    "SIM",
    "I",
    "C",
    "RUF",
]

[tool.ty.src]
exclude = ["**/*.ipynb", "scripts/"]

[dependency-groups]
dev = [
    "bandit",
    "pytest",
    "ruff",
    "ty",
    "types-requests",
    "pandas-stubs",
]
script = [
    "jupyterlab",
    "mapclassify",
    "matplotlib",
    "folium",
    "ipympl",
    "scipy",
    "streamlit",
]
